requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
jupyter==1.0.0
